A model-based clustering

A model-based clustering based on parameterized finite Gaussian mixture models

data(EMexample)

Format

an object of class Mclust providing the optimal (according to BIC) mixture model estimation composed of 5 clusters. The details of the output components are as follows:

call the matched call

data a matrix, the input data matrix

modelName a character string denoting the model at which the optimal BIC occurs

n an integer, the number of observations in the data

d a double, the dimension of the data

G an integer, the optimal number of mixture components

BIC a double, all BIC values

bic a double, the optimal BIC value

loglik a double, the log-likelihood corresponding to the optimal BIC

df a double, the number of estimated parameters

hypvol NA

parameters a list with the following components:

pro a vector of double whose kth component is the mixing proportion for the kth component of the mixture model

mean a matrix of double whose kth column is the mean of the kth component of the mixture model

variance a list of variance parameters for the model

z a matrix of double whose [i,k]th entry is the probability that observation i in the test data belongs to the kth class

classification an array of double, the classification corresponding to z

uncertainty a double, the uncertainty associated with the classification

Value

an object of class Mclust providing the optimal (according to BIC) mixture model estimation composed of 5 clusters. The details of the output components are as follows:

call the matched call

data a matrix, the input data matrix

modelName a character string denoting the model at which the optimal BIC occurs

n an integer, the number of observations in the data

d a double, the dimension of the data

G an integer, the optimal number of mixture components

BIC a double, all BIC values

bic a double, the optimal BIC value

loglik a double, the log-likelihood corresponding to the optimal BIC

df a double, the number of estimated parameters

hypvol NA

parameters a list with the following components:

pro a vector of double whose kth component is the mixing proportion for the kth component of the mixture model

mean a matrix of double whose kth column is the mean of the kth component of the mixture model

variance a list of variance parameters for the model

z a matrix of double whose [i,k]th entry is the probability that observation i in the test data belongs to the kth class

classification an array of double, the classification corresponding to z

uncertainty a double, the uncertainty associated with the classification

Details

See Mclust for detailed description of the object of class mclust.

See also

Examples

## Loading the demo object of class 'Mclust'
data(EMexample)

## Group clusters that have at least 2% of overlap
## The inital object has 5 clusters while the return object has only 
## 4 clusters
CNprep:::consolidate(EMexample, minover=0.2)

#> $mu
#>           1           2           3           4           5 
#> -0.23327452 -0.08138647 -0.02177458  0.03224144  0.24586682 
#> 
#> $pro
#> [1] 0.09326425 0.29018379 0.23315391 0.22536697 0.15803109
#> 
#> $z
#>                 [,1]          [,2]          [,3]          [,4]          [,5]
#>   [1,] 1.518326e-120  2.430697e-24  6.977993e-07  9.999993e-01  3.216325e-67
#>   [2,] 3.595740e-126  9.516728e-27  5.003598e-08  9.999999e-01  1.082023e-62
#>   [3,] 1.575231e-121  9.218454e-25  4.400933e-07  9.999996e-01  1.991578e-66
#>   [4,] 2.685068e-133  8.488695e-30  1.775920e-09  1.000000e+00  5.877146e-57
#>   [5,] 8.480066e-123  2.640361e-25  2.428818e-07  9.999998e-01  2.090796e-65
#>   [6,] 3.060337e-122  4.572648e-25  3.153439e-07  9.999997e-01  7.443863e-66
#>   [7,] 3.187284e-118  2.395315e-23  2.070719e-06  9.999979e-01  4.353225e-69
#>   [8,]  1.000000e+00  3.496168e-23  2.767048e-52  2.524535e-88  0.000000e+00
#>   [9,]  1.000000e+00  6.425580e-24  2.616218e-53  1.306859e-89  0.000000e+00
#>  [10,]  1.000000e+00  4.436161e-32  1.134045e-64 7.098136e-104  0.000000e+00
#>  [11,]  1.000000e+00  3.620265e-32  8.545341e-65 4.975715e-104  0.000000e+00
#>  [12,]  1.000000e+00  1.228463e-26  4.289744e-57  2.313461e-94  0.000000e+00
#>  [13,]  1.000000e+00  3.571670e-30  5.108677e-62 1.522485e-100  0.000000e+00
#>  [14,]  1.000000e+00  5.952846e-26  3.861121e-56  3.649732e-93  0.000000e+00
#>  [15,]  1.000000e+00  1.886382e-30  2.100312e-62 4.988162e-101  0.000000e+00
#>  [16,]  1.000000e+00  3.498971e-25  4.547266e-55  8.069252e-92  0.000000e+00
#>  [17,]  1.000000e+00  2.736723e-28  2.148049e-59  2.994832e-97  0.000000e+00
#>  [18,]  1.000000e+00  1.319747e-31  5.174836e-64 4.772904e-103  0.000000e+00
#>  [19,]  1.000000e+00  4.232806e-32  1.062319e-64 6.539164e-104  0.000000e+00
#>  [20,]  1.000000e+00  7.019656e-30  1.308861e-61 4.960070e-100  0.000000e+00
#>  [21,] 6.074535e-120  4.399429e-24  9.251868e-07  9.999991e-01  1.053935e-67
#>  [22,] 8.595568e-121  1.905453e-24  6.215317e-07  9.999994e-01  5.083848e-67
#>  [23,] 8.595568e-121  1.905453e-24  6.215317e-07  9.999994e-01  5.083848e-67
#>  [24,] 8.114711e-120  4.979782e-24  9.813267e-07  9.999990e-01  8.348634e-68
#>  [25,] 1.005949e-123  1.060459e-25  1.574159e-07  9.999998e-01  1.162298e-64
#>  [26,] 7.594533e-125  3.510316e-26  9.306309e-08  9.999999e-01  9.294927e-64
#>  [27,] 8.541281e-123  2.648500e-25  2.432374e-07  9.999998e-01  2.078729e-65
#>  [28,] 1.678264e-117  4.876116e-23  2.903275e-06  9.999971e-01  1.143593e-69
#>  [29,] 1.504708e-122  3.374704e-25  2.729364e-07  9.999997e-01  1.317952e-65
#>  [30,] 3.218946e-121  1.251608e-24  5.089619e-07  9.999995e-01  1.120583e-66
#>  [31,] 2.091672e-122  3.885479e-25  2.918511e-07  9.999997e-01  1.011102e-65
#>  [32,] 1.500923e-123  1.258511e-25  1.707663e-07  9.999998e-01  8.423168e-65
#>  [33,] 2.254461e-124  5.591801e-26  1.161207e-07  9.999999e-01  3.872522e-64
#>  [34,] 7.112440e-119  1.260727e-23  1.526168e-06  9.999985e-01  1.455432e-68
#>  [35,] 3.836090e-122  5.036793e-25  3.301760e-07  9.999997e-01  6.206432e-66
#>  [36,] 1.504708e-122  3.374704e-25  2.729364e-07  9.999997e-01  1.317952e-65
#>  [37,] 4.948763e-124  7.828227e-26  1.362618e-07  9.999999e-01  2.056967e-64
#>  [38,] 8.541281e-123  2.648500e-25  2.432374e-07  9.999998e-01  2.078729e-65
#>  [39,] 1.678264e-117  4.876116e-23  2.903275e-06  9.999971e-01  1.143593e-69
#>  [40,] 7.484915e-127  4.862093e-27  3.635965e-08  1.000000e+00  3.825762e-62
#>  [41,] 2.813406e-125  2.295101e-26  7.603991e-08  9.999999e-01  2.066741e-63
#>  [42,] 3.075372e-119  8.806631e-24  1.286847e-06  9.999987e-01  2.857581e-68
#>  [43,] 5.259733e-120  4.136479e-24  8.984721e-07  9.999991e-01  1.183442e-67
#>  [44,] 3.801921e-118  2.583051e-23  2.146352e-06  9.999979e-01  3.777336e-69
#>  [45,] 2.091672e-122  3.885479e-25  2.918511e-07  9.999997e-01  1.011102e-65
#>  [46,] 1.194917e-125  1.590978e-26  6.388290e-08  9.999999e-01  4.116687e-63
#>  [47,]  1.000000e+00  3.748095e-28  3.328240e-59  5.189327e-97  0.000000e+00
#>  [48,]  1.000000e+00  1.498617e-34  4.107990e-68 3.398884e-108  0.000000e+00
#>  [49,]  1.000000e+00  5.817924e-38  7.317193e-73 3.707813e-114  0.000000e+00
#>  [50,]  1.000000e+00  9.356656e-30  1.952848e-61 8.196791e-100  0.000000e+00
#>  [51,]  1.000000e+00  5.709093e-32  1.611308e-64 1.103195e-103  0.000000e+00
#>  [52,]  1.000000e+00  8.698358e-31  7.147891e-63 1.289092e-101  0.000000e+00
#>  [53,]  1.000000e+00  1.590717e-33  1.101719e-66 2.111502e-106  0.000000e+00
#>  [54,]  1.000000e+00  2.248514e-23  1.496590e-52  1.167078e-88  0.000000e+00
#>  [55,]  1.000000e+00  4.076632e-28  3.741310e-59  6.010304e-97  0.000000e+00
#>  [56,]  1.000000e+00  3.527140e-26  1.863046e-56  1.461982e-93  0.000000e+00
#>  [57,]  1.000000e+00  5.155694e-36  3.767232e-70 9.404568e-111  0.000000e+00
#>  [58,]  1.000000e+00  2.573032e-35  3.532812e-69 1.562083e-109  0.000000e+00
#>  [59,]  1.000000e+00  5.806512e-35  1.097195e-68 6.479809e-109  0.000000e+00
#>  [60,]  1.000000e+00  6.543596e-23  6.623004e-52  7.551329e-88  0.000000e+00
#>  [61,] 1.467280e-114  8.847813e-22  1.151662e-05  9.999885e-01  4.910550e-72
#>  [62,] 9.141573e-113  5.184668e-21  2.669246e-05  9.999733e-01  1.766387e-73
#>  [63,] 1.366585e-115  3.204160e-22  7.105729e-06  9.999929e-01  3.316463e-71
#>  [64,] 2.296091e-124  5.635754e-26  1.165538e-07  9.999999e-01  3.815922e-64
#>  [65,] 1.591224e-117  4.766251e-23  2.871990e-06  9.999971e-01  1.193667e-69
#>  [66,] 1.106904e-113  2.100717e-21  1.737235e-05  9.999826e-01  9.658889e-73
#>  [67,] 1.391110e-112  6.205058e-21  2.907251e-05  9.999709e-01  1.259956e-73
#>  [68,] 1.141207e-112  5.700947e-21  2.792466e-05  9.999721e-01  1.477603e-73
#>  [69,] 9.520600e-115  7.352816e-22  1.054654e-05  9.999895e-01  6.954916e-72
#>  [70,] 2.387992e-117  5.670457e-23  3.119242e-06  9.999969e-01  8.610213e-70
#>  [71,]  0.000000e+00 8.283441e-172 2.193768e-115  6.467252e-74  1.000000e+00
#>  [72,]  0.000000e+00 6.346946e-173 2.683605e-116  1.208873e-74  1.000000e+00
#>  [73,]  0.000000e+00 1.291930e-168 8.968633e-113  7.857038e-72  1.000000e+00
#>  [74,]  0.000000e+00 7.185108e-170 8.441843e-114  1.191435e-72  1.000000e+00
#>  [75,]  0.000000e+00 9.342815e-176 1.295365e-118  1.711901e-76  1.000000e+00
#>  [76,]  0.000000e+00 4.331409e-176 6.907834e-119  1.036403e-76  1.000000e+00
#>  [77,]  0.000000e+00 9.231175e-170 1.036198e-113  1.403185e-72  1.000000e+00
#>  [78,]  0.000000e+00 1.448733e-167 6.475931e-112  3.806957e-71  1.000000e+00
#>  [79,]  0.000000e+00 2.817379e-167 1.115718e-111  5.876999e-71  1.000000e+00
#>  [80,]  0.000000e+00 4.375713e-171 8.558357e-115  1.916981e-73  1.000000e+00
#>  [81,]  0.000000e+00 7.331663e-173 3.019599e-116  1.328229e-74  1.000000e+00
#>  [82,]  0.000000e+00 1.494710e-172 5.407120e-116  2.114619e-74  1.000000e+00
#>  [83,]  0.000000e+00 8.988854e-167 2.881736e-111  1.253416e-70  1.000000e+00
#>  [84,]  0.000000e+00 1.343972e-175 1.744001e-118  2.170553e-76  1.000000e+00
#>  [85,]  0.000000e+00 7.907577e-170 9.129975e-114  1.268339e-72  1.000000e+00
#>  [86,]  0.000000e+00 5.924443e-169 4.740201e-113  4.722941e-72  1.000000e+00
#>  [87,]  0.000000e+00 2.817379e-167 1.115718e-111  5.876999e-71  1.000000e+00
#>  [88,]  0.000000e+00 1.378949e-165 2.688686e-110  7.451823e-70  1.000000e+00
#>  [89,]  0.000000e+00 6.309418e-168 3.281300e-112  2.212601e-71  1.000000e+00
#>  [90,]  0.000000e+00 1.378949e-165 2.688686e-110  7.451823e-70  1.000000e+00
#>  [91,]  0.000000e+00 2.792448e-175 3.171782e-118  3.498720e-76  1.000000e+00
#>  [92,]  0.000000e+00 1.291930e-168 8.968633e-113  7.857038e-72  1.000000e+00
#>  [93,]  0.000000e+00 7.317298e-172 1.982163e-115  5.964276e-74  1.000000e+00
#>  [94,]  0.000000e+00 5.043903e-170 6.320557e-114  9.456940e-73  1.000000e+00
#>  [95,]  0.000000e+00 3.132632e-170 4.281239e-114  6.929561e-73  1.000000e+00
#>  [96,]  0.000000e+00 8.060812e-169 6.097808e-113  5.774544e-72  1.000000e+00
#>  [97,]  0.000000e+00 1.080066e-169 1.178195e-113  1.554656e-72  1.000000e+00
#>  [98,]  0.000000e+00 7.291707e-175 6.953999e-118  6.546995e-76  1.000000e+00
#>  [99,]  0.000000e+00 8.988854e-167 2.881736e-111  1.253416e-70  1.000000e+00
#> [100,]  0.000000e+00 7.039123e-177 1.562927e-119  3.164937e-77  1.000000e+00
#> [101,]  0.000000e+00 3.543831e-175 3.854293e-118  4.087618e-76  1.000000e+00
#> [102,]  0.000000e+00 2.410948e-173 1.215900e-116  6.426008e-75  1.000000e+00
#> [103,]  0.000000e+00 3.184665e-171 6.599853e-115  1.557886e-73  1.000000e+00
#> [104,]  0.000000e+00 1.326825e-175 1.725781e-118  2.152433e-76  1.000000e+00
#> [105,]  0.000000e+00 6.496252e-171 1.182364e-114  2.481159e-73  1.000000e+00
#> [106,]  0.000000e+00 6.484529e-166 1.450586e-110  4.553511e-70  1.000000e+00
#> [107,]  0.000000e+00 6.407168e-164 6.209460e-109  9.133246e-69  1.000000e+00
#> [108,]  0.000000e+00 6.496252e-171 1.182364e-114  2.481159e-73  1.000000e+00
#> [109,]  0.000000e+00 7.440912e-167 2.469014e-111  1.107928e-70  1.000000e+00
#> [110,]  0.000000e+00 1.068310e-173 6.248599e-117  3.777185e-75  1.000000e+00
#> [111,]  0.000000e+00 2.308750e-166 6.233328e-111  2.320312e-70  1.000000e+00
#> [112,]  0.000000e+00 3.958594e-171 7.885076e-115  1.795619e-73  1.000000e+00
#> [113,]  0.000000e+00 2.820920e-171 5.976594e-115  1.439291e-73  1.000000e+00
#> [114,]  0.000000e+00 8.283441e-172 2.193768e-115  6.467252e-74  1.000000e+00
#> [115,]  0.000000e+00 1.911025e-166 5.340265e-111  2.050885e-70  1.000000e+00
#> [116,]  0.000000e+00 1.240386e-175 1.633266e-118  2.059822e-76  1.000000e+00
#> [117,]  0.000000e+00 8.283441e-172 2.193768e-115  6.467252e-74  1.000000e+00
#> [118,]  0.000000e+00 1.279341e-171 3.130312e-115  8.589345e-74  1.000000e+00
#> [119,]  0.000000e+00 7.331663e-173 3.019599e-116  1.328229e-74  1.000000e+00
#> [120,]  0.000000e+00 6.465973e-169 5.091729e-113  5.000478e-72  1.000000e+00
#> [121,]  0.000000e+00 1.800991e-168 1.176860e-112  9.759911e-72  1.000000e+00
#> [122,]  0.000000e+00 2.095404e-174 1.648866e-117  1.304156e-75  1.000000e+00
#> [123,]  0.000000e+00 9.691452e-166 2.014991e-110  5.919348e-70  1.000000e+00
#> [124,]  0.000000e+00 3.389203e-167 1.297752e-111  6.630520e-71  1.000000e+00
#> [125,]  0.000000e+00 5.217176e-171 9.882482e-115  2.150237e-73  1.000000e+00
#> [126,]  0.000000e+00 6.484529e-166 1.450586e-110  4.553511e-70  1.000000e+00
#> [127,]  0.000000e+00 6.465973e-169 5.091729e-113  5.000478e-72  1.000000e+00
#> [128,]  0.000000e+00 4.259299e-173 1.936595e-116  9.317323e-75  1.000000e+00
#> [129,]  0.000000e+00 7.331663e-173 3.019599e-116  1.328229e-74  1.000000e+00
#> [130,]  0.000000e+00 2.308750e-166 6.233328e-111  2.320312e-70  1.000000e+00
#> [131,]  0.000000e+00 3.954119e-169 3.405474e-113  3.627230e-72  1.000000e+00
#> [132,]  4.369643e-68  3.892733e-05  9.999589e-01  2.142088e-06 1.155221e-119
#> [133,]  2.530440e-64  4.471794e-04  9.995526e-01  2.342265e-07 1.995822e-124
#> [134,]  9.666337e-68  4.868799e-05  9.999496e-01  1.748915e-06 4.229181e-120
#> [135,]  3.117329e-63  9.071283e-04  9.990927e-01  1.232699e-07 8.306549e-126
#> [136,]  7.462601e-68  4.526434e-05  9.999529e-01  1.868398e-06 5.867828e-120
#> [137,]  4.103997e-66  1.400082e-04  9.998593e-01  6.713575e-07 3.679909e-122
#> [138,]  2.755606e-68  3.418463e-05  9.999634e-01  2.409777e-06 2.070521e-119
#> [139,]  5.268137e-68  4.103359e-05  9.999569e-01  2.042187e-06 9.117637e-120
#> [140,]  5.268137e-68  4.103359e-05  9.999569e-01  2.042187e-06 9.117637e-120
#> [141,]  1.263721e-65  1.922127e-04  9.998073e-01  5.037097e-07 8.863600e-123
#> [142,]  3.718600e-66  1.361713e-04  9.998631e-01  6.884839e-07 4.169112e-122
#> [143,]  3.376563e-66  1.325189e-04  9.998668e-01  7.056639e-07 4.710659e-122
#> [144,]  2.755606e-68  3.418463e-05  9.999634e-01  2.409777e-06 2.070521e-119
#> [145,]  4.994915e-65  2.831101e-04  9.997165e-01  3.545624e-07 1.556368e-123
#> [146,]  1.220112e-65  1.903202e-04  9.998092e-01  5.082489e-07 9.266468e-123
#> [147,]  2.150715e-65  2.232799e-04  9.997763e-01  4.397290e-07 4.521808e-123
#> [148,]  2.150715e-65  2.232799e-04  9.997763e-01  4.397290e-07 4.521808e-123
#> [149,]  1.292112e-66  1.010943e-04  9.998980e-01  9.019000e-07 1.588874e-121
#> [150,]  5.279335e-66  1.503043e-04  9.998491e-01  6.295285e-07 2.675503e-122
#> [151,]  3.620436e-65  2.585689e-04  9.997410e-01  3.849476e-07 2.338977e-123
#> [152,]  3.718600e-66  1.361713e-04  9.998631e-01  6.884839e-07 4.169112e-122
#> [153,]  5.268137e-68  4.103359e-05  9.999569e-01  2.042187e-06 9.117637e-120
#> [154,]  1.342037e-64  3.740134e-04  9.996257e-01  2.754351e-07 4.454313e-124
#> [155,]  4.367095e-69  2.034162e-05  9.999758e-01  3.857484e-06 2.131102e-118
#> [156,]  4.384099e-65  2.728947e-04  9.997267e-01  3.665774e-07 1.835760e-123
#> [157,]  4.370816e-64  5.216138e-04  9.994782e-01  2.036919e-07 9.991649e-125
#> [158,]  1.484489e-68  2.871633e-05  9.999685e-01  2.822204e-06 4.529769e-119
#> [159,]  2.277346e-65  2.269082e-04  9.997727e-01  4.333488e-07 4.205940e-123
#> [160,]  1.743493e-66  1.099998e-04  9.998892e-01  8.354527e-07 1.087428e-121
#> [161,]  2.918474e-66  1.271850e-04  9.998721e-01  7.324400e-07 5.665309e-122
#> [162,]  5.084797e-64  5.443266e-04  9.994555e-01  1.959653e-07 8.249938e-125
#> [163,]  1.530043e-69  1.513661e-05  9.999798e-01  5.042373e-06 8.036666e-118
#> [164,]  7.462601e-68  4.526434e-05  9.999529e-01  1.868398e-06 5.867828e-120
#> [165,]  5.220215e-68  4.092806e-05  9.999570e-01  2.046959e-06 9.223699e-120
#> [166,]  5.084797e-64  5.443266e-04  9.994555e-01  1.959653e-07 8.249938e-125
#> [167,]  1.457641e-65  2.001019e-04  9.997994e-01  4.856706e-07 7.398365e-123
#> [168,]  3.376563e-66  1.325189e-04  9.998668e-01  7.056639e-07 4.710659e-122
#> [169,]  3.868201e-68  3.761293e-05  9.999602e-01  2.209824e-06 1.347918e-119
#> [170,]  2.044620e-65  2.201199e-04  9.997794e-01  4.454489e-07 4.820817e-123
#> [171,]  4.384099e-65  2.728947e-04  9.997267e-01  3.665774e-07 1.835760e-123
#> [172,]  1.274315e-64  3.685969e-04  9.996311e-01  2.791036e-07 4.756051e-124
#> [173,]  9.248738e-67  9.200398e-05  9.999070e-01  9.823152e-07 2.425974e-121
#> [174,]  9.666337e-68  4.868799e-05  9.999496e-01  1.748915e-06 4.229181e-120
#> [175,]  2.093603e-65  2.215931e-04  9.997780e-01  4.427628e-07 4.678500e-123
#> [176,]  2.734770e-65  2.389173e-04  9.997607e-01  4.135513e-07 3.336156e-123
#> [177,]  2.179954e-67  6.122752e-05  9.999374e-01  1.420892e-06 1.510957e-120
#> [178,]  1.827860e-65  2.132781e-04  9.997863e-01  4.583862e-07 5.555511e-123
#> [179,]  3.393356e-68  3.625006e-05  9.999615e-01  2.284994e-06 1.590932e-119
#> [180,]  9.644386e-68  4.865680e-05  9.999496e-01  1.749931e-06 4.241368e-120
#> [181,]  1.889495e-69  1.606399e-05  9.999792e-01  4.777821e-06 6.153086e-118
#> [182,]  1.831686e-66  1.115400e-04  9.998876e-01  8.249882e-07 1.021589e-121
#> [183,]  1.056499e-63  6.688367e-04  9.993310e-01  1.625548e-07 3.268727e-125
#> [184,]  2.406718e-67  6.295895e-05  9.999357e-01  1.385427e-06 1.333084e-120
#> [185,]  6.279060e-65  3.019613e-04  9.996977e-01  3.344298e-07 1.165025e-123
#> [186,]  2.831578e-68  3.444763e-05  9.999632e-01  2.393096e-06 2.000464e-119
#> [187,]  5.183941e-66  1.495341e-04  9.998498e-01  6.324676e-07 2.737970e-122
#> [188,]  2.277346e-65  2.269082e-04  9.997727e-01  4.333488e-07 4.205940e-123
#> [189,]  1.827860e-65  2.132781e-04  9.997863e-01  4.583862e-07 5.555511e-123
#> [190,]  5.545518e-66  1.524021e-04  9.998470e-01  6.216677e-07 2.514007e-122
#> [191,]  3.620436e-65  2.585689e-04  9.997410e-01  3.849476e-07 2.338977e-123
#> [192,]  1.467051e-66  1.047769e-04  9.998943e-01  8.731176e-07 1.352988e-121
#> [193,]  1.268130e-84  8.300913e-10  9.656969e-01  3.430311e-02  9.077704e-99
#> [194,]  1.268130e-84  8.300913e-10  9.656969e-01  3.430311e-02  9.077704e-99
#> [195,]  1.452199e-90  1.099598e-11  5.026512e-01  4.973488e-01  6.833210e-92
#> [196,]  1.347569e-84  8.447443e-10  9.662029e-01  3.379714e-02  8.415800e-99
#> [197,]  1.347569e-84  8.447443e-10  9.662029e-01  3.379714e-02  8.415800e-99
#> [198,]  4.621315e-84  1.203349e-09  9.750450e-01  2.495496e-02  1.805887e-99
#> [199,]  3.800254e-83  2.195324e-09  9.852385e-01  1.476150e-02 1.284808e-100
#> [200,]  7.903682e-84  1.403017e-09  9.781541e-01  2.184588e-02 9.222675e-100
#>  [ reached getOption("max.print") -- omitted 186 rows ]
#> 
#> $groups
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    0    0    0    0
#> [2,]    0    1    0    0    0
#> [3,]    0    0    1    0    0
#> [4,]    0    0    0    1    0
#> [5,]    0    0    0    0    1
#> 
#> $ngroups
#> [1] 5
#> 
#> $sigmasq
#> [1] 0.0001367895 0.0001367895 0.0001367895 0.0001367895 0.0001367895
#>