Annotation table for ROMA CGH platform and human genome version 17.

Whole genome annotation table using Representational Oligonucleotide Microarray Analysis (ROMA) CGH platform, human genome version 17.

data(annotexample)

Format

a data.frame with 83055 observations on the following 3 variables.

PROBEID

a character vector of probe names.

CHROM

a numeric vector, used as integer, of chromosome positions.

CHROM.POS

a numeric vector, used as integer, of genomic positions.

Source

GEO accession GPL9775, http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GPL9775

Value

a data.frame with 83055 observations on the following 3 variables.

PROBEID

a character vector of probe names.

CHROM

a numeric vector, used as integer, of chromosome positions.

CHROM.POS

a numeric vector, used as integer, of genomic positions.

Details

The values in the chromosome column are all integer, with 23 corresponding to X, 24 to Y and 25 to a set of non-human test probes.

See also

Examples

## Loading annotation table dataset
data(annotexample)

## Loading other datasets
data(ratexample)
data(segexample)
data(normsegs)

## How to use annotexample, when segment table does not have columns 
## of integer postions in terms of measuring units(probes), such 
## as "mysegs" below 

mysegs <- segexample[,c(1,5:12)]

## Analysis limited to chromosomes 1 and 2
## The bsTimes variable should be higher for a real analysis
segtable <- CNpreprocessing(segall=mysegs, ratall=ratexample, idCol="ID",
    chromCol="chrom", bpStartCol="chrom.pos.start", bpEndCol="chrom.pos.end",
    annot=annotexample, annotStartCol="CHROM.POS", annotEndCol="CHROM.POS",
    annotChromCol = "CHROM", blsize = 50, minJoin = 0.25, cWeight = 0.4, 
    bsTimes = 3, chromRange = 1:2,  nJobs = 1, modelNames = "E", 
    normalLength = normsegs[,1], normalMedian = normsegs[,2])