Example of a boundary positions table.

A table of genomic positions for DNA copy-number changing events, collected from genomes of 1203 individuals using Representational Oligonucleotide Microarray Analysis (ROMA) platform.

data(cnpexample)

Format

a data.frame with 19188 rows and 4 columns:

copy.num

a character vector indicating whether an event is a gain ("amp") or a loss ("del").

chrom

a numeric vector used as an integer indicating which chromosome the event is in.

chrom.start

a numeric vector, used as an integer, representing the event start positions.

chrom.end

a numeric vector, used as an integer, of event start positions.

Source

Strong association of de novo copy number mutations with autism. Sebat J, Lakshmi B, Malhotra D, Troge J, Lese-Martin C, Walsh T, Yamrom B, Yoon S, Krasnitz A, Kendall J, Leotta A, Pai D, Zhang R, Lee YH, Hicks J, Bregman J, Sutcliffe JS, Jobanputra V, Chung W, Warburton D, King MC, Skuse D, Geschwind DH, Gilliam TC, Ye K, Wigler M. Science. 2007 Apr 20;316(5823):445-9.

Value

a data.frame with 19188 rows and 4 columns:

copy.num

a character vector indicating whether an event is a gain ("amp") or a loss ("del").

chrom

a numeric vector used as an integer indicating which chromosome the event is in.

chrom.start

a numeric vector, used as an integer, representing the event start positions.

chrom.end

a numeric vector, used as an integer, of event start positions.

See also

Examples

## Load datasets
data(segexample)
data(ratexample)
data(normsegs)
data(cnpexample)

## Create a table with segment information for profile WZ2 with help
## of normal samples
segtable <- CNpreprocessing(segall = segexample[segexample[,"ID"] == "WZ2",],
    ratall = ratexample, idCol = "ID", startCol = "start", endCol = "end",
    chromCol = "chrom", bpStartCol = "chrom.pos.start", 
    bpEndCol = "chrom.pos.end", blsize = 50, minJoin = 0.25, cWeight = 0.4,
    bsTimes = 50, chromRange = 1:22, nJobs = 1, 
    modelNames = "E", normalLength = normsegs[,1], 
    normalMedian = normsegs[,2])

## Add an eventIndex column to segtable that identifies the 
## amplication (marked as 1) and deletion (marked as -1) events
eventIndex <- rep(0, nrow(segtable))
eventIndex[segtable[,"marginalprob"] < 1e-4 & segtable[,"negtail"] > 0.999 & 
    segtable[,"mediandev"] < 0] <- -1
eventIndex[segtable[,"marginalprob"] < 1e-4 & segtable[,"negtail"] > 0.999 &
    segtable[,"mediandev"] > 0] <- 1
segtable <- cbind(segtable, eventIndex)

## Create a mask table using amplification and deletion regions as input
namps17 <- cnpexample[cnpexample[,"copy.num"] == "amp",]
aCNPmask <- makeCNPmask(imat=namps17, chromCol=2, startCol=3, 
    endCol=4, nProf=1203, uThresh=0.02, dThresh=0.008)
ndels17 <- cnpexample[cnpexample[,"copy.num"] == "del",]
dCNPmask <- makeCNPmask(imat=ndels17, chromCol=2, startCol=3, 
    endCol=4, nProf=1203, uThresh=0.02, dThresh=0.008)
maskTable <- rbind(cbind(aCNPmask, cnpindex=1), 
    cbind(dCNPmask, cnpindex=-1))

## Apply a mask to a table of copy number events
maskedCNP <- applyCNPmask(segTable=segtable, chrom="chrom",
    startPos="chrom.pos.start", endPos="chrom.pos.end", 
    startProbe="start", endProbe="end", eventIndex="eventIndex",
    maskTable=maskTable, maskChrom="chrom", maskStart="start", 
    maskEnd="end", maskIndex="cnpindex", minCover=0.005,
    indexVals=c(-1, 1))

## Show some results
tail(maskedCNP)

#>        StartProbe EndProbe toremove
#> [165,]      80389    81843        0
#> [166,]      81844    81854        0
#> [167,]      81855    81877        0
#> [168,]      81878    82039        0
#> [169,]      82040    82055        0
#> [170,]      82056    83055        0