TODO

TODO

getz(logr, emfit, zgroup, times)

Arguments

logr
A vector of double representing the log ratio values.

emfit
An object of class Mclust providing a mixture model estimation.

zgroup
A matrix of double, used as integer, with the number of rows corresponding to the current number of clusters while the number of columns is corresponding to the initial number of clusters. The presence of 1 in position [i,k] indicates that the initial ith cluster is now part of the new kth cluster.

times
A single double value specifying the number of time the median of each segment is sampled in order to predict the cluster assignment for the segment.

Value

A matrix of double containing the probability for each observation to belong to each class of the mixture model with the updated number of clusters.

Author

Alexander Krasnitz, Guoli Sun

Examples

# TODO

## Load 
data(EMexample)