A reference set of segments

A table of segment lengths and log copy number ratios for a large set of human diploid genomes.

data(normsegs)

Format

a matrix with 43497 rows/segments and 2 columns/variables. The 2 columns are:

length a numeric vector, used as integer, of segment genomic length

segmedian a numeric vector of segment median computed from log copy number ratio

Source

Sebat J, et al. Strong association of de novo copy number mutations with autism. Science. 2007 Apr 20;316(5823):445-9. Epub 2007 Mar 15.

Value

a matrix with 43497 rows/segments and 2 columns/variables. The 2 columns are:

length a numeric vector, used as integer, of segment genomic length

segmedian a numeric vector of segment median computed from log copy number ratio

Details

The table originates in a set of copy number profiles of over 1000 individuals, obtained using Representational Oligonucleotide Microarray Analysis (ROMA) technology. To ensure ploidy of 2 segments from X and Y chromosomes and segments shorter than 5Mb were excluded.

See also

Examples

## Loading log ratio dataset
data(segexample)

## Load datasets
data(segexample)
data(ratexample)
data(normsegs)

## Preprocess segments for WZ3 sample
segtable <- CNpreprocessing(segall=segexample[segexample[,"ID"] == "WZ3",],
                 ratall = ratexample, idCol = "ID", startCol = "start", 
                 endCol="end", chromCol = "chrom", 
                 bpStartCol = "chrom.pos.start", 
                 bpEndCol = "chrom.pos.end", blsize = 50, 
                 minJoin = 0.30, cWeight = 0.4,
                 bsTimes = 40, chromRange = 1:22, 
                 modelNames = "E", normalLength = normsegs[,1], 
                 normalMedian = normsegs[,2])