Example of a segmented copy number table

Segmented log ratio data for 5 breast cancer genomes, derived using Representational Oligonucleotide Microarray Analysis (ROMA) platform. ROMA detects genomic amplifications and deletions with boundaries defined at a resolution of 50 kb. In this segmented table, each row represents a segment.

data(segexample)

Format

a data.frame with 479 rows/segments and 12 columns/variables:

ID a character vector of profile IDs

start a numeric vector (segment start probe number)

end a numeric vector (segment end probe number)

num.probes a numeric vector (number of probes in the segment)

seg.median a numeric vector (median log ratio)

chrom a numeric vector (chromosome number)

chrom.pos.start a numeric vector (genomic start)

chrom.pos.end a numeric vector (genomic end)

cytoband.start a character vector (cytogenetic band start)

cytoband.end a character vector (cytogenetic band end)

abs.pos.start a numeric vector (genomic start, absolute)

abs.pos.end a numeric vector (genomic end, absolute)

Source

Hicks, J. et al. Novel patterns of genome rearrangement and their association with survival in breast cancer. Genome Res. 2006. 16:1465–1479. doi: 10.1101/gr.5460106

Value

a data.frame with 479 rows/segments and 12 columns/variables:

ID a character vector of profile IDs

start a numeric vector (segment start probe number)

end a numeric vector (segment end probe number)

num.probes a numeric vector (number of probes in the segment)

seg.median a numeric vector (median log ratio)

chrom a numeric vector (chromosome number)

chrom.pos.start a numeric vector (genomic start)

chrom.pos.end a numeric vector (genomic end)

cytoband.start a character vector (cytogenetic band start)

cytoband.end a character vector (cytogenetic band end)

abs.pos.start a numeric vector (genomic start, absolute)

abs.pos.end a numeric vector (genomic end, absolute)

Details

Segment medians are computed from log copy number ratio. The corresponding raw data table is ratexample in this package.

See also

Examples

## Loading log ratio dataset
data(segexample)

## Load datasets
data(segexample)
data(ratexample)
data(normsegs)

## Preprocess segments for WZ2 sample
segtable <- CNpreprocessing(segall=segexample[segexample[,"ID"] == "WZ2",],
                 ratall = ratexample, idCol = "ID", startCol = "start", 
                 endCol = "end", chromCol = "chrom", 
                 bpStartCol = "chrom.pos.start", 
                 bpEndCol = "chrom.pos.end", blsize = 50, 
                 minJoin = 0.25, cWeight = 0.4,
                 bsTimes = 30, chromRange = 1:22, nJobs = 1,
                 modelNames = "E", normalLength = normsegs[,1], 
                 normalMedian = normsegs[,2])