Parameters validation for the makeCNPmask function

Validation of all parameters needed by the public makeCNPmask function.

validateMakeCNPmask(imat, chromCol, startCol, endCol, nProf, uThresh, dThresh)

Arguments

imat
A matrix or a data.frame tabulating the chromosome numbers and startpoint/endpoint positions of the interval events.

chromCol
A character string or numeric representing an integer specifying the column of imat containing the chromosome numbers of the interval events.

startCol
A character string or numeric representing an integer specifying the column of imat containing the left (right) endpoint of the interval events.

endCol
A character string or numeric representing an integer specifying the column of imat containing the right endpoint of the interval events.

nProf
A positive numeric acting as an integer specifying the number of copy number profiles from which the events originated.

uThresh
A numeric specifying the upper ratio threshold for the event frequency or (if nProf = 1) for the event count. The upper ratio threshold must be equal or superior to the lower threshold (dThresh). The upper ratio threshold must also be between 0 and 1.

dThresh
A numeric specifying the lower ratio threshold for the event frequency or (if nProf = 1) for the event count. The lower ratio threshold must be equal or inferior to the upper threshold (dThresh). The lower ratio threshold must also be between 0 and 1.

Value

0.

Author

Astrid Deschênes

Examples

data(segexample)

## Return zero as all parameters are valid
CNprep:::makeCNPmask(imat=segexample, chromCol="chrom",
    startCol="start", endCol="end", nProf=1203, uThresh=0.20, dThresh=0.10)

#>      chrom start end