A model-based clustering

A model-based clustering based on parameterized finite Gaussian mixture models

data(EMexample)

Format

an object of class Mclust providing the optimal (according to BIC) mixture model estimation composed of 5 clusters. The details of the output components are as follows:

call the matched call

data a matrix, the input data matrix

modelName a character string denoting the model at which the optimal BIC occurs

n an integer, the number of observations in the data

d a double, the dimension of the data

G an integer, the optimal number of mixture components

BIC a double, all BIC values

bic a double, the optimal BIC value

loglik a double, the log-likelihood corresponding to the optimal BIC

df a double, the number of estimated parameters

hypvol NA

parameters a list with the following components:

pro a vector of double whose kth component is the mixing proportion for the kth component of the mixture model

mean a matrix of double whose kth column is the mean of the kth component of the mixture model

variance a list of variance parameters for the model

z a matrix of double whose [i,k]th entry is the probability that observation i in the test data belongs to the kth class

classification an array of double, the classification corresponding to z

uncertainty a double, the uncertainty associated with the classification

Value

an object of class Mclust providing the optimal (according to BIC) mixture model estimation composed of 5 clusters. The details of the output components are as follows:

call the matched call

data a matrix, the input data matrix

modelName a character string denoting the model at which the optimal BIC occurs

n an integer, the number of observations in the data

d a double, the dimension of the data

G an integer, the optimal number of mixture components

BIC a double, all BIC values

bic a double, the optimal BIC value

loglik a double, the log-likelihood corresponding to the optimal BIC

df a double, the number of estimated parameters

hypvol NA

parameters a list with the following components:

pro a vector of double whose kth component is the mixing proportion for the kth component of the mixture model

mean a matrix of double whose kth column is the mean of the kth component of the mixture model

variance a list of variance parameters for the model

z a matrix of double whose [i,k]th entry is the probability that observation i in the test data belongs to the kth class

classification an array of double, the classification corresponding to z

uncertainty a double, the uncertainty associated with the classification

Details

See Mclust for detailed description of the object of class mclust.

See also

Examples

## Loading the demo object of class 'Mclust'
data(EMexample)

## Group clusters that have at least 2% of overlap
## The inital object has 5 clusters while the return object has only 
## 4 clusters
CNprep:::consolidate(EMexample, minover=0.2)

#> $mu
#>           1           2           3           4           5 
#> -0.23327452 -0.08138647 -0.02177458  0.03224144  0.24586682 
#> 
#> $pro
#> [1] 0.09326425 0.29018379 0.23315391 0.22536697 0.15803109
#> 
#> $z
#>                 [,1]          [,2]          [,3]          [,4]          [,5]
#>   [1,] 1.518326e-120  2.430697e-24  6.977993e-07  9.999993e-01  3.216325e-67
#>   [2,] 3.595740e-126  9.516728e-27  5.003598e-08  9.999999e-01  1.082023e-62
#>   [3,] 1.575231e-121  9.218454e-25  4.400933e-07  9.999996e-01  1.991578e-66
#>   [4,] 2.685068e-133  8.488695e-30  1.775920e-09  1.000000e+00  5.877146e-57
#>   [5,] 8.480066e-123  2.640361e-25  2.428818e-07  9.999998e-01  2.090796e-65
#>   [6,] 3.060337e-122  4.572648e-25  3.153439e-07  9.999997e-01  7.443863e-66
#>   [7,] 3.187284e-118  2.395315e-23  2.070719e-06  9.999979e-01  4.353225e-69
#>   [8,]  1.000000e+00  3.496168e-23  2.767048e-52  2.524535e-88  0.000000e+00
#>   [9,]  1.000000e+00  6.425580e-24  2.616218e-53  1.306859e-89  0.000000e+00
#>  [10,]  1.000000e+00  4.436161e-32  1.134045e-64 7.098136e-104  0.000000e+00
#>  [11,]  1.000000e+00  3.620265e-32  8.545341e-65 4.975715e-104  0.000000e+00
#>  [12,]  1.000000e+00  1.228463e-26  4.289744e-57  2.313461e-94  0.000000e+00
#>  [13,]  1.000000e+00  3.571670e-30  5.108677e-62 1.522485e-100  0.000000e+00
#>  [14,]  1.000000e+00  5.952846e-26  3.861121e-56  3.649732e-93  0.000000e+00
#>  [15,]  1.000000e+00  1.886382e-30  2.100312e-62 4.988162e-101  0.000000e+00
#>  [16,]  1.000000e+00  3.498971e-25  4.547266e-55  8.069252e-92  0.000000e+00
#>  [17,]  1.000000e+00  2.736723e-28  2.148049e-59  2.994832e-97  0.000000e+00
#>  [18,]  1.000000e+00  1.319747e-31  5.174836e-64 4.772904e-103  0.000000e+00
#>  [19,]  1.000000e+00  4.232806e-32  1.062319e-64 6.539164e-104  0.000000e+00
#>  [20,]  1.000000e+00  7.019656e-30  1.308861e-61 4.960070e-100  0.000000e+00
#>  [21,] 6.074535e-120  4.399429e-24  9.251868e-07  9.999991e-01  1.053935e-67
#>  [22,] 8.595568e-121  1.905453e-24  6.215317e-07  9.999994e-01  5.083848e-67
#>  [23,] 8.595568e-121  1.905453e-24  6.215317e-07  9.999994e-01  5.083848e-67
#>  [24,] 8.114711e-120  4.979782e-24  9.813267e-07  9.999990e-01  8.348634e-68
#>  [25,] 1.005949e-123  1.060459e-25  1.574159e-07  9.999998e-01  1.162298e-64
#>  [26,] 7.594533e-125  3.510316e-26  9.306309e-08  9.999999e-01  9.294927e-64
#>  [27,] 8.541281e-123  2.648500e-25  2.432374e-07  9.999998e-01  2.078729e-65
#>  [28,] 1.678264e-117  4.876116e-23  2.903275e-06  9.999971e-01  1.143593e-69
#>  [29,] 1.504708e-122  3.374704e-25  2.729364e-07  9.999997e-01  1.317952e-65
#>  [30,] 3.218946e-121  1.251608e-24  5.089619e-07  9.999995e-01  1.120583e-66
#>  [31,] 2.091672e-122  3.885479e-25  2.918511e-07  9.999997e-01  1.011102e-65
#>  [32,] 1.500923e-123  1.258511e-25  1.707663e-07  9.999998e-01  8.423168e-65
#>  [33,] 2.254461e-124  5.591801e-26  1.161207e-07  9.999999e-01  3.872522e-64
#>  [34,] 7.112440e-119  1.260727e-23  1.526168e-06  9.999985e-01  1.455432e-68
#>  [35,] 3.836090e-122  5.036793e-25  3.301760e-07  9.999997e-01  6.206432e-66
#>  [36,] 1.504708e-122  3.374704e-25  2.729364e-07  9.999997e-01  1.317952e-65
#>  [37,] 4.948763e-124  7.828227e-26  1.362618e-07  9.999999e-01  2.056967e-64
#>  [38,] 8.541281e-123  2.648500e-25  2.432374e-07  9.999998e-01  2.078729e-65
#>  [39,] 1.678264e-117  4.876116e-23  2.903275e-06  9.999971e-01  1.143593e-69
#>  [40,] 7.484915e-127  4.862093e-27  3.635965e-08  1.000000e+00  3.825762e-62
#>  [41,] 2.813406e-125  2.295101e-26  7.603991e-08  9.999999e-01  2.066741e-63
#>  [42,] 3.075372e-119  8.806631e-24  1.286847e-06  9.999987e-01  2.857581e-68
#>  [43,] 5.259733e-120  4.136479e-24  8.984721e-07  9.999991e-01  1.183442e-67
#>  [44,] 3.801921e-118  2.583051e-23  2.146352e-06  9.999979e-01  3.777336e-69
#>  [45,] 2.091672e-122  3.885479e-25  2.918511e-07  9.999997e-01  1.011102e-65
#>  [46,] 1.194917e-125  1.590978e-26  6.388290e-08  9.999999e-01  4.116687e-63
#>  [47,]  1.000000e+00  3.748095e-28  3.328240e-59  5.189327e-97  0.000000e+00
#>  [48,]  1.000000e+00  1.498617e-34  4.107990e-68 3.398884e-108  0.000000e+00
#>  [49,]  1.000000e+00  5.817924e-38  7.317193e-73 3.707813e-114  0.000000e+00
#>  [50,]  1.000000e+00  9.356656e-30  1.952848e-61 8.196791e-100  0.000000e+00
#>  [51,]  1.000000e+00  5.709093e-32  1.611308e-64 1.103195e-103  0.000000e+00
#>  [52,]  1.000000e+00  8.698358e-31  7.147891e-63 1.289092e-101  0.000000e+00
#>  [53,]  1.000000e+00  1.590717e-33  1.101719e-66 2.111502e-106  0.000000e+00
#>  [54,]  1.000000e+00  2.248514e-23  1.496590e-52  1.167078e-88  0.000000e+00
#>  [55,]  1.000000e+00  4.076632e-28  3.741310e-59  6.010304e-97  0.000000e+00
#>  [56,]  1.000000e+00  3.527140e-26  1.863046e-56  1.461982e-93  0.000000e+00
#>  [57,]  1.000000e+00  5.155694e-36  3.767232e-70 9.404568e-111  0.000000e+00
#>  [58,]  1.000000e+00  2.573032e-35  3.532812e-69 1.562083e-109  0.000000e+00
#>  [59,]  1.000000e+00  5.806512e-35  1.097195e-68 6.479809e-109  0.000000e+00
#>  [60,]  1.000000e+00  6.543596e-23  6.623004e-52  7.551329e-88  0.000000e+00
#>  [61,] 1.467280e-114  8.847813e-22  1.151662e-05  9.999885e-01  4.910550e-72
#>  [62,] 9.141573e-113  5.184668e-21  2.669246e-05  9.999733e-01  1.766387e-73
#>  [63,] 1.366585e-115  3.204160e-22  7.105729e-06  9.999929e-01  3.316463e-71
#>  [64,] 2.296091e-124  5.635754e-26  1.165538e-07  9.999999e-01  3.815922e-64
#>  [65,] 1.591224e-117  4.766251e-23  2.871990e-06  9.999971e-01  1.193667e-69
#>  [66,] 1.106904e-113  2.100717e-21  1.737235e-05  9.999826e-01  9.658889e-73
#>  [67,] 1.391110e-112  6.205058e-21  2.907251e-05  9.999709e-01  1.259956e-73
#>  [68,] 1.141207e-112  5.700947e-21  2.792466e-05  9.999721e-01  1.477603e-73
#>  [69,] 9.520600e-115  7.352816e-22  1.054654e-05  9.999895e-01  6.954916e-72
#>  [70,] 2.387992e-117  5.670457e-23  3.119242e-06  9.999969e-01  8.610213e-70
#>  [71,]  0.000000e+00 8.283441e-172 2.193768e-115  6.467252e-74  1.000000e+00
#>  [72,]  0.000000e+00 6.346946e-173 2.683605e-116  1.208873e-74  1.000000e+00
#>  [73,]  0.000000e+00 1.291930e-168 8.968633e-113  7.857038e-72  1.000000e+00
#>  [74,]  0.000000e+00 7.185108e-170 8.441843e-114  1.191435e-72  1.000000e+00
#>  [75,]  0.000000e+00 9.342815e-176 1.295365e-118  1.711901e-76  1.000000e+00
#>  [76,]  0.000000e+00 4.331409e-176 6.907834e-119  1.036403e-76  1.000000e+00
#>  [77,]  0.000000e+00 9.231175e-170 1.036198e-113  1.403185e-72  1.000000e+00
#>  [78,]  0.000000e+00 1.448733e-167 6.475931e-112  3.806957e-71  1.000000e+00
#>  [79,]  0.000000e+00 2.817379e-167 1.115718e-111  5.876999e-71  1.000000e+00
#>  [80,]  0.000000e+00 4.375713e-171 8.558357e-115  1.916981e-73  1.000000e+00
#>  [81,]  0.000000e+00 7.331663e-173 3.019599e-116  1.328229e-74  1.000000e+00
#>  [82,]  0.000000e+00 1.494710e-172 5.407120e-116  2.114619e-74  1.000000e+00
#>  [83,]  0.000000e+00 8.988854e-167 2.881736e-111  1.253416e-70  1.000000e+00
#>  [84,]  0.000000e+00 1.343972e-175 1.744001e-118  2.170553e-76  1.000000e+00
#>  [85,]  0.000000e+00 7.907577e-170 9.129975e-114  1.268339e-72  1.000000e+00
#>  [86,]  0.000000e+00 5.924443e-169 4.740201e-113  4.722941e-72  1.000000e+00
#>  [87,]  0.000000e+00 2.817379e-167 1.115718e-111  5.876999e-71  1.000000e+00
#>  [88,]  0.000000e+00 1.378949e-165 2.688686e-110  7.451823e-70  1.000000e+00
#>  [89,]  0.000000e+00 6.309418e-168 3.281300e-112  2.212601e-71  1.000000e+00
#>  [90,]  0.000000e+00 1.378949e-165 2.688686e-110  7.451823e-70  1.000000e+00
#>  [91,]  0.000000e+00 2.792448e-175 3.171782e-118  3.498720e-76  1.000000e+00
#>  [92,]  0.000000e+00 1.291930e-168 8.968633e-113  7.857038e-72  1.000000e+00
#>  [93,]  0.000000e+00 7.317298e-172 1.982163e-115  5.964276e-74  1.000000e+00
#>  [94,]  0.000000e+00 5.043903e-170 6.320557e-114  9.456940e-73  1.000000e+00
#>  [95,]  0.000000e+00 3.132632e-170 4.281239e-114  6.929561e-73  1.000000e+00
#>  [96,]  0.000000e+00 8.060812e-169 6.097808e-113  5.774544e-72  1.000000e+00
#>  [97,]  0.000000e+00 1.080066e-169 1.178195e-113  1.554656e-72  1.000000e+00
#>  [98,]  0.000000e+00 7.291707e-175 6.953999e-118  6.546995e-76  1.000000e+00
#>  [99,]  0.000000e+00 8.988854e-167 2.881736e-111  1.253416e-70  1.000000e+00
#> [100,]  0.000000e+00 7.039123e-177 1.562927e-119  3.164937e-77  1.000000e+00
#> [101,]  0.000000e+00 3.543831e-175 3.854293e-118  4.087618e-76  1.000000e+00
#> [102,]  0.000000e+00 2.410948e-173 1.215900e-116  6.426008e-75  1.000000e+00
#> [103,]  0.000000e+00 3.184665e-171 6.599853e-115  1.557886e-73  1.000000e+00
#> [104,]  0.000000e+00 1.326825e-175 1.725781e-118  2.152433e-76  1.000000e+00
#> [105,]  0.000000e+00 6.496252e-171 1.182364e-114  2.481159e-73  1.000000e+00
#> [106,]  0.000000e+00 6.484529e-166 1.450586e-110  4.553511e-70  1.000000e+00
#> [107,]  0.000000e+00 6.407168e-164 6.209460e-109  9.133246e-69  1.000000e+00
#> [108,]  0.000000e+00 6.496252e-171 1.182364e-114  2.481159e-73  1.000000e+00
#> [109,]  0.000000e+00 7.440912e-167 2.469014e-111  1.107928e-70  1.000000e+00
#> [110,]  0.000000e+00 1.068310e-173 6.248599e-117  3.777185e-75  1.000000e+00
#> [111,]  0.000000e+00 2.308750e-166 6.233328e-111  2.320312e-70  1.000000e+00
#> [112,]  0.000000e+00 3.958594e-171 7.885076e-115  1.795619e-73  1.000000e+00
#> [113,]  0.000000e+00 2.820920e-171 5.976594e-115  1.439291e-73  1.000000e+00
#> [114,]  0.000000e+00 8.283441e-172 2.193768e-115  6.467252e-74  1.000000e+00
#> [115,]  0.000000e+00 1.911025e-166 5.340265e-111  2.050885e-70  1.000000e+00
#> [116,]  0.000000e+00 1.240386e-175 1.633266e-118  2.059822e-76  1.000000e+00
#> [117,]  0.000000e+00 8.283441e-172 2.193768e-115  6.467252e-74  1.000000e+00
#> [118,]  0.000000e+00 1.279341e-171 3.130312e-115  8.589345e-74  1.000000e+00
#> [119,]  0.000000e+00 7.331663e-173 3.019599e-116  1.328229e-74  1.000000e+00
#> [120,]  0.000000e+00 6.465973e-169 5.091729e-113  5.000478e-72  1.000000e+00
#> [121,]  0.000000e+00 1.800991e-168 1.176860e-112  9.759911e-72  1.000000e+00
#> [122,]  0.000000e+00 2.095404e-174 1.648866e-117  1.304156e-75  1.000000e+00
#> [123,]  0.000000e+00 9.691452e-166 2.014991e-110  5.919348e-70  1.000000e+00
#> [124,]  0.000000e+00 3.389203e-167 1.297752e-111  6.630520e-71  1.000000e+00
#> [125,]  0.000000e+00 5.217176e-171 9.882482e-115  2.150237e-73  1.000000e+00
#> [126,]  0.000000e+00 6.484529e-166 1.450586e-110  4.553511e-70  1.000000e+00
#> [127,]  0.000000e+00 6.465973e-169 5.091729e-113  5.000478e-72  1.000000e+00
#> [128,]  0.000000e+00 4.259299e-173 1.936595e-116  9.317323e-75  1.000000e+00
#> [129,]  0.000000e+00 7.331663e-173 3.019599e-116  1.328229e-74  1.000000e+00
#> [130,]  0.000000e+00 2.308750e-166 6.233328e-111  2.320312e-70  1.000000e+00
#> [131,]  0.000000e+00 3.954119e-169 3.405474e-113  3.627230e-72  1.000000e+00
#> [132,]  4.369643e-68  3.892733e-05  9.999589e-01  2.142088e-06 1.155221e-119
#> [133,]  2.530440e-64  4.471794e-04  9.995526e-01  2.342265e-07 1.995822e-124
#> [134,]  9.666337e-68  4.868799e-05  9.999496e-01  1.748915e-06 4.229181e-120
#> [135,]  3.117329e-63  9.071283e-04  9.990927e-01  1.232699e-07 8.306549e-126
#> [136,]  7.462601e-68  4.526434e-05  9.999529e-01  1.868398e-06 5.867828e-120
#> [137,]  4.103997e-66  1.400082e-04  9.998593e-01  6.713575e-07 3.679909e-122
#> [138,]  2.755606e-68  3.418463e-05  9.999634e-01  2.409777e-06 2.070521e-119
#> [139,]  5.268137e-68  4.103359e-05  9.999569e-01  2.042187e-06 9.117637e-120
#> [140,]  5.268137e-68  4.103359e-05  9.999569e-01  2.042187e-06 9.117637e-120
#> [141,]  1.263721e-65  1.922127e-04  9.998073e-01  5.037097e-07 8.863600e-123
#> [142,]  3.718600e-66  1.361713e-04  9.998631e-01  6.884839e-07 4.169112e-122
#> [143,]  3.376563e-66  1.325189e-04  9.998668e-01  7.056639e-07 4.710659e-122
#> [144,]  2.755606e-68  3.418463e-05  9.999634e-01  2.409777e-06 2.070521e-119
#> [145,]  4.994915e-65  2.831101e-04  9.997165e-01  3.545624e-07 1.556368e-123
#> [146,]  1.220112e-65  1.903202e-04  9.998092e-01  5.082489e-07 9.266468e-123
#> [147,]  2.150715e-65  2.232799e-04  9.997763e-01  4.397290e-07 4.521808e-123
#> [148,]  2.150715e-65  2.232799e-04  9.997763e-01  4.397290e-07 4.521808e-123
#> [149,]  1.292112e-66  1.010943e-04  9.998980e-01  9.019000e-07 1.588874e-121
#> [150,]  5.279335e-66  1.503043e-04  9.998491e-01  6.295285e-07 2.675503e-122
#> [151,]  3.620436e-65  2.585689e-04  9.997410e-01  3.849476e-07 2.338977e-123
#> [152,]  3.718600e-66  1.361713e-04  9.998631e-01  6.884839e-07 4.169112e-122
#> [153,]  5.268137e-68  4.103359e-05  9.999569e-01  2.042187e-06 9.117637e-120
#> [154,]  1.342037e-64  3.740134e-04  9.996257e-01  2.754351e-07 4.454313e-124
#> [155,]  4.367095e-69  2.034162e-05  9.999758e-01  3.857484e-06 2.131102e-118
#> [156,]  4.384099e-65  2.728947e-04  9.997267e-01  3.665774e-07 1.835760e-123
#> [157,]  4.370816e-64  5.216138e-04  9.994782e-01  2.036919e-07 9.991649e-125
#> [158,]  1.484489e-68  2.871633e-05  9.999685e-01  2.822204e-06 4.529769e-119
#> [159,]  2.277346e-65  2.269082e-04  9.997727e-01  4.333488e-07 4.205940e-123
#> [160,]  1.743493e-66  1.099998e-04  9.998892e-01  8.354527e-07 1.087428e-121
#> [161,]  2.918474e-66  1.271850e-04  9.998721e-01  7.324400e-07 5.665309e-122
#> [162,]  5.084797e-64  5.443266e-04  9.994555e-01  1.959653e-07 8.249938e-125
#> [163,]  1.530043e-69  1.513661e-05  9.999798e-01  5.042373e-06 8.036666e-118
#> [164,]  7.462601e-68  4.526434e-05  9.999529e-01  1.868398e-06 5.867828e-120
#> [165,]  5.220215e-68  4.092806e-05  9.999570e-01  2.046959e-06 9.223699e-120
#> [166,]  5.084797e-64  5.443266e-04  9.994555e-01  1.959653e-07 8.249938e-125
#> [167,]  1.457641e-65  2.001019e-04  9.997994e-01  4.856706e-07 7.398365e-123
#> [168,]  3.376563e-66  1.325189e-04  9.998668e-01  7.056639e-07 4.710659e-122
#> [169,]  3.868201e-68  3.761293e-05  9.999602e-01  2.209824e-06 1.347918e-119
#> [170,]  2.044620e-65  2.201199e-04  9.997794e-01  4.454489e-07 4.820817e-123
#> [171,]  4.384099e-65  2.728947e-04  9.997267e-01  3.665774e-07 1.835760e-123
#> [172,]  1.274315e-64  3.685969e-04  9.996311e-01  2.791036e-07 4.756051e-124
#> [173,]  9.248738e-67  9.200398e-05  9.999070e-01  9.823152e-07 2.425974e-121
#> [174,]  9.666337e-68  4.868799e-05  9.999496e-01  1.748915e-06 4.229181e-120
#> [175,]  2.093603e-65  2.215931e-04  9.997780e-01  4.427628e-07 4.678500e-123
#> [176,]  2.734770e-65  2.389173e-04  9.997607e-01  4.135513e-07 3.336156e-123
#> [177,]  2.179954e-67  6.122752e-05  9.999374e-01  1.420892e-06 1.510957e-120
#> [178,]  1.827860e-65  2.132781e-04  9.997863e-01  4.583862e-07 5.555511e-123
#> [179,]  3.393356e-68  3.625006e-05  9.999615e-01  2.284994e-06 1.590932e-119
#> [180,]  9.644386e-68  4.865680e-05  9.999496e-01  1.749931e-06 4.241368e-120
#> [181,]  1.889495e-69  1.606399e-05  9.999792e-01  4.777821e-06 6.153086e-118
#> [182,]  1.831686e-66  1.115400e-04  9.998876e-01  8.249882e-07 1.021589e-121
#> [183,]  1.056499e-63  6.688367e-04  9.993310e-01  1.625548e-07 3.268727e-125
#> [184,]  2.406718e-67  6.295895e-05  9.999357e-01  1.385427e-06 1.333084e-120
#> [185,]  6.279060e-65  3.019613e-04  9.996977e-01  3.344298e-07 1.165025e-123
#> [186,]  2.831578e-68  3.444763e-05  9.999632e-01  2.393096e-06 2.000464e-119
#> [187,]  5.183941e-66  1.495341e-04  9.998498e-01  6.324676e-07 2.737970e-122
#> [188,]  2.277346e-65  2.269082e-04  9.997727e-01  4.333488e-07 4.205940e-123
#> [189,]  1.827860e-65  2.132781e-04  9.997863e-01  4.583862e-07 5.555511e-123
#> [190,]  5.545518e-66  1.524021e-04  9.998470e-01  6.216677e-07 2.514007e-122
#> [191,]  3.620436e-65  2.585689e-04  9.997410e-01  3.849476e-07 2.338977e-123
#> [192,]  1.467051e-66  1.047769e-04  9.998943e-01  8.731176e-07 1.352988e-121
#> [193,]  1.268130e-84  8.300913e-10  9.656969e-01  3.430311e-02  9.077704e-99
#> [194,]  1.268130e-84  8.300913e-10  9.656969e-01  3.430311e-02  9.077704e-99
#> [195,]  1.452199e-90  1.099598e-11  5.026512e-01  4.973488e-01  6.833210e-92
#> [196,]  1.347569e-84  8.447443e-10  9.662029e-01  3.379714e-02  8.415800e-99
#> [197,]  1.347569e-84  8.447443e-10  9.662029e-01  3.379714e-02  8.415800e-99
#> [198,]  4.621315e-84  1.203349e-09  9.750450e-01  2.495496e-02  1.805887e-99
#> [199,]  3.800254e-83  2.195324e-09  9.852385e-01  1.476150e-02 1.284808e-100
#> [200,]  7.903682e-84  1.403017e-09  9.781541e-01  2.184588e-02 9.222675e-100
#> [201,]  5.593962e-82  4.708793e-09  9.925035e-01  7.496540e-03 4.344722e-102
#> [202,]  1.783820e-88  5.722856e-11  7.567202e-01  2.432798e-01  3.916493e-94
#> [203,]  2.467169e-86  2.626240e-10  9.126076e-01  8.739236e-02  1.168800e-96
#> [204,]  3.631416e-88  7.191354e-11  7.869006e-01  2.130994e-01  1.740423e-94
#> [205,]  5.235316e-86  3.281901e-10  9.265038e-01  7.349616e-02  4.667475e-97
#> [206,]  9.938561e-86  3.962778e-10  9.367360e-01  6.326397e-02  2.126033e-97
#> [207,]  5.142054e-88  8.031672e-11  8.007093e-01  1.992907e-01  1.165705e-94
#> [208,]  2.399234e-80  1.362563e-08  9.971127e-01  2.887325e-03 3.772348e-104
#> [209,]  1.347569e-84  8.447443e-10  9.662029e-01  3.379714e-02  8.415800e-99
#> [210,]  1.562516e-85  4.523808e-10  9.431547e-01  5.684531e-02  1.217860e-97
#> [211,]  1.347569e-84  8.447443e-10  9.662029e-01  3.379714e-02  8.415800e-99
#> [212,]  5.593962e-82  4.708793e-09  9.925035e-01  7.496540e-03 4.344722e-102
#> [213,]  4.756398e-84  1.213318e-09  9.752223e-01  2.477766e-02  1.741941e-99
#> [214,]  4.745796e-79  3.163154e-08  9.986501e-01  1.349852e-03 8.661807e-106
#> [215,]  1.035232e-82  2.918667e-09  9.885240e-01  1.147604e-02 3.641601e-101
#> [216,]  1.000000e+00  6.606254e-60 2.036694e-103 1.622042e-152  0.000000e+00
#> [217,]  1.000000e+00  4.879277e-61 5.411455e-105 1.706280e-154  0.000000e+00
#> [218,]  1.000000e+00  6.606254e-60 2.036694e-103 1.622042e-152  0.000000e+00
#> [219,]  1.000000e+00  1.749200e-57 4.813826e-100 2.787475e-148  0.000000e+00
#> [220,]  1.000000e+00  3.670668e-63 5.976557e-108 3.311505e-158  0.000000e+00
#> [221,]  1.000000e+00  2.946826e-62 1.086443e-106 1.262592e-156  0.000000e+00
#> [222,]  4.186924e-71  5.490670e-06  9.999819e-01  1.264056e-05 7.637855e-116
#> [223,]  1.000000e+00  7.589947e-45  1.898869e-82 3.428548e-126  0.000000e+00
#> [224,]  1.107239e-89  2.258404e-11  6.170476e-01  3.829524e-01  8.314726e-93
#> [225,]  1.739927e-86  2.366580e-10  9.054011e-01  9.459887e-02  1.786038e-96
#> [226,]  2.301382e-91  5.561213e-12  4.008197e-01  5.991803e-01  4.211190e-91
#> [227,]  3.712437e-89  3.410759e-11  6.814936e-01  3.185064e-01  2.252283e-93
#> [228,]  4.814686e-94  4.844545e-13  1.508295e-01  8.491705e-01  1.131933e-88
#> [229,]  5.591864e-87  1.681616e-10  8.783170e-01  1.216830e-01  7.013222e-96
#> [230,]  2.100543e-94  3.444708e-13  1.299040e-01  8.700960e-01  2.305584e-88
#> [231,]  6.899423e-82  4.996912e-09  9.928910e-01  7.109010e-03 3.334712e-102
#> [232,]  2.712247e-82  3.835660e-09  9.909983e-01  9.001739e-03 1.082334e-101
#> [233,]  7.856415e-81  9.940959e-09  9.961647e-01  3.835339e-03 1.546324e-103
#> [234,]  1.000000e+00  4.952239e-63 9.068564e-108 5.589350e-158  0.000000e+00
#> [235,]  1.000000e+00  1.485520e-49  5.280679e-89 2.018362e-134  0.000000e+00
#> [236,] 3.761667e-109  1.825214e-19  1.450911e-04  9.998549e-01  2.180392e-76
#> [237,] 1.756439e-105  6.780483e-18  8.088562e-04  9.991911e-01  2.428752e-79
#> [238,] 4.192788e-108  5.121130e-19  2.369365e-04  9.997631e-01  3.132188e-77
#> [239,] 7.161270e-106  4.619148e-18  6.739852e-04  9.993260e-01  5.000724e-79
#> [240,] 5.438087e-110  7.978274e-20  9.790021e-05  9.999021e-01  1.033862e-75
#> [241,] 2.471519e-105  7.847269e-18  8.670161e-04  9.991330e-01  1.844915e-79
#> [242,] 1.461947e-110  4.547634e-20  7.494248e-05  9.999251e-01  2.975576e-75
#> [243,] 1.013985e-108  2.789883e-19  1.775169e-04  9.998225e-01  9.816715e-77
#> [244,] 2.431365e-107  1.086388e-18  3.387580e-04  9.996612e-01  7.612126e-78
#> [245,] 4.262542e-105  9.907902e-18  9.686043e-04  9.990314e-01  1.189656e-79
#> [246,] 4.262542e-105  9.907902e-18  9.686043e-04  9.990314e-01  1.189656e-79
#> [247,] 2.471519e-105  7.847269e-18  8.670161e-04  9.991330e-01  1.844915e-79
#> [248,] 9.550994e-107  1.950840e-18  4.474388e-04  9.995526e-01  2.530882e-78
#> [249,] 9.550994e-107  1.950840e-18  4.474388e-04  9.995526e-01  2.530882e-78
#> [250,] 7.161270e-106  4.619148e-18  6.739852e-04  9.993260e-01  5.000724e-79
#> [251,] 1.112379e-104  1.493450e-17  1.177125e-03  9.988229e-01  5.495845e-80
#> [252,] 5.436183e-111  2.978122e-20  6.128140e-05  9.999387e-01  6.596435e-75
#> [253,] 7.161270e-106  4.619148e-18  6.739852e-04  9.993260e-01  5.000724e-79
#> [254,] 5.371677e-106  4.084453e-18  6.357091e-04  9.993643e-01  6.303088e-79
#> [255,] 2.008257e-104  1.922830e-17  1.327290e-03  9.986727e-01  3.415534e-80
#> [256,] 1.461947e-110  4.547634e-20  7.494248e-05  9.999251e-01  2.975576e-75
#> [257,] 1.013985e-108  2.789883e-19  1.775169e-04  9.998225e-01  9.816715e-77
#> [258,] 1.166837e-106  2.125342e-18  4.660349e-04  9.995340e-01  2.154168e-78
#> [259,] 1.756439e-105  6.780483e-18  8.088562e-04  9.991911e-01  2.428752e-79
#> [260,] 9.740136e-109  2.742283e-19  1.760706e-04  9.998239e-01  1.013964e-76
#> [261,] 3.472859e-107  1.265420e-18  3.642338e-04  9.996358e-01  5.713331e-78
#> [262,] 3.780903e-110  6.829120e-20  9.092293e-05  9.999091e-01  1.385128e-75
#> [263,] 1.756439e-105  6.780483e-18  8.088562e-04  9.991911e-01  2.428752e-79
#> [264,] 5.593369e-105  1.112919e-17  1.023608e-03  9.989764e-01  9.559174e-80
#> [265,] 1.583396e-107  9.042509e-19  3.104610e-04  9.996895e-01  1.075007e-77
#> [266,] 5.436183e-111  2.978122e-20  6.128140e-05  9.999387e-01  6.596435e-75
#> [267,] 5.341685e-107  1.521395e-18  3.975663e-04  9.996024e-01  4.040102e-78
#> [268,] 1.461947e-110  4.547634e-20  7.494248e-05  9.999251e-01  2.975576e-75
#> [269,] 6.016379e-110  8.330844e-20  9.993360e-05  9.999001e-01  9.531118e-76
#> [270,] 3.761667e-109  1.825214e-19  1.450911e-04  9.998549e-01  2.180392e-76
#> [271,] 6.813198e-107  1.688323e-18  4.177344e-04  9.995823e-01  3.321560e-78
#> [272,] 2.046512e-109  1.406673e-19  1.281931e-04  9.998718e-01  3.558622e-76
#> [273,] 7.495186e-104  3.377446e-17  1.734534e-03  9.982655e-01  1.182723e-80
#> [274,] 4.192788e-108  5.121130e-19  2.369365e-04  9.997631e-01  3.132188e-77
#> [275,]  3.272035e-37  9.999989e-01  1.060857e-06  8.823004e-22 2.835052e-172
#> [276,]  1.078011e-36  9.999993e-01  6.644819e-07  3.616677e-22 2.172274e-173
#> [277,]  8.398456e-38  9.999982e-01  1.808842e-06  2.439977e-21 5.309917e-171
#> [278,]  1.048662e-37  9.999983e-01  1.657913e-06  2.066587e-21 3.290865e-171
#> [279,]  7.371203e-38  9.999981e-01  1.903852e-06  2.690090e-21 7.033441e-171
#> [280,]  2.233282e-36  9.999995e-01  4.993062e-07  2.097534e-22 4.522526e-174
#> [281,]  2.487209e-36  9.999995e-01  4.786477e-07  1.935205e-22 3.586035e-174
#> [282,]  3.628252e-38  9.999975e-01  2.514328e-06  4.571142e-21 3.239151e-170
#> [283,]  6.156720e-39  9.999950e-01  5.042966e-06  1.722766e-20 1.479774e-168
#> [284,]  3.762826e-37  9.999990e-01  1.004249e-06  7.947253e-22 2.097910e-172
#> [285,]  1.255505e-38  9.999962e-01  3.812929e-06  1.011000e-20 3.187313e-169
#> [286,]  2.771410e-36  9.999995e-01  4.587530e-07  1.784764e-22 2.840369e-174
#> [287,]  3.180413e-38  9.999974e-01  2.647715e-06  5.044515e-21 4.302318e-170
#> [288,]  8.398456e-38  9.999982e-01  1.808842e-06  2.439977e-21 5.309917e-171
#> [289,]  8.174763e-37  9.999993e-01  7.406725e-07  4.448133e-22 3.942590e-173
#> [290,]  8.041095e-38  9.999982e-01  1.839969e-06  2.520639e-21 5.831432e-171
#> [291,]  1.103165e-38  9.999960e-01  4.011451e-06  1.113707e-20 4.211760e-169
#> [292,]  4.911658e-38  9.999978e-01  2.232609e-06  3.644541e-21 1.686708e-170
#> [293,]  2.195281e-36  9.999995e-01  5.026798e-07  2.124634e-22 4.692886e-174
#> [294,]  3.536741e-38  9.999975e-01  2.539657e-06  4.659324e-21 3.422430e-170
#> [295,]  1.078011e-36  9.999993e-01  6.644819e-07  3.616677e-22 2.172274e-173
#> [296,]  1.110729e-37  9.999984e-01  1.620926e-06  1.979588e-21 2.907402e-171
#> [297,]  5.236649e-38  9.999978e-01  2.177182e-06  3.474003e-21 1.469225e-170
#> [298,]  6.545433e-39  9.999951e-01  4.923266e-06  1.645654e-20 1.296904e-168
#> [299,]  4.925850e-39  9.999945e-01  5.504204e-06  2.035548e-20 2.392787e-168
#> [300,]  8.505936e-38  9.999982e-01  1.799840e-06  2.416879e-21 5.166409e-171
#> [301,]  3.412952e-38  9.999974e-01  2.575409e-06  4.785159e-21 3.695491e-170
#> [302,]  3.032785e-37  9.999989e-01  1.092939e-06  9.338605e-22 3.338955e-172
#> [303,]  7.953619e-36  9.999997e-01  3.033363e-07  8.111637e-23 2.930053e-175
#> [304,]  5.475081e-38  9.999979e-01  2.139476e-06  3.360210e-21 1.334828e-170
#> [305,]  5.039078e-37  9.999991e-01  8.955167e-07  6.387720e-22 1.118170e-172
#> [306,]  2.481567e-36  9.999995e-01  4.790744e-07  1.938495e-22 3.603623e-174
#> [307,]  3.679694e-38  9.999975e-01  2.500477e-06  4.523258e-21 3.142373e-170
#> [308,]  3.791342e-38  9.999975e-01  2.471322e-06  4.423252e-21 2.946380e-170
#> [309,]  4.517285e-39  9.999943e-01  5.694416e-06  2.171741e-20 2.883524e-168
#> [310,]  4.911658e-38  9.999978e-01  2.232609e-06  3.644541e-21 1.686708e-170
#> [311,]  2.224873e-37  9.999988e-01  1.234198e-06  1.177368e-21 6.508458e-172
#> [312,]  1.393621e-36  9.999994e-01  6.007930e-07  2.984654e-22 1.249205e-173
#> [313,]  9.103727e-38  9.999982e-01  1.752507e-06  2.297162e-21 4.463087e-171
#> [314,]  3.855384e-38  9.999975e-01  2.455133e-06  4.368179e-21 2.841941e-170
#> [315,]  6.792155e-38  9.999980e-01  1.965959e-06  2.859849e-21 8.389212e-171
#> [316,]  1.289139e-37  9.999985e-01  1.528902e-06  1.770871e-21 2.109222e-171
#> [317,]  1.794113e-37  9.999987e-01  1.342933e-06  1.382976e-21 1.034747e-171
#> [318,]  1.794113e-37  9.999987e-01  1.342933e-06  1.382976e-21 1.034747e-171
#> [319,]  2.295246e-36  9.999995e-01  4.939730e-07  2.055033e-22 4.263558e-174
#> [320,]  2.012050e-39  9.999922e-01  7.821055e-06  3.976722e-20 1.646980e-167
#> [321,]  3.272035e-37  9.999989e-01  1.060857e-06  8.823004e-22 2.835052e-172
#> [322,]  8.026829e-39  9.999955e-01  4.544508e-06  1.412745e-20 8.356056e-169
#> [323,]  1.363173e-38  9.999963e-01  3.691800e-06  9.506574e-21 2.669541e-169
#> [324,]  1.771327e-39  9.999918e-01  8.222033e-06  4.374397e-20 2.167314e-167
#> [325,]  3.412952e-38  9.999974e-01  2.575409e-06  4.785159e-21 3.695491e-170
#> [326,]  3.615296e-37  9.999990e-01  1.020133e-06  8.188599e-22 2.286716e-172
#> [327,]  7.430984e-40  9.999884e-01  1.156117e-05  8.377071e-20 1.408423e-166
#> [328,]  9.103727e-38  9.999982e-01  1.752507e-06  2.297162e-21 4.463087e-171
#> [329,]  3.032785e-37  9.999989e-01  1.092939e-06  9.338605e-22 3.338955e-172
#> [330,]  1.030332e-36  9.999993e-01  6.763815e-07  3.741144e-22 2.394655e-173
#> [331,]  6.822481e-37  9.999992e-01  7.951344e-07  5.092353e-22 5.820842e-173
#> [332,]  4.911658e-38  9.999978e-01  2.232609e-06  3.644541e-21 1.686708e-170
#> [333,]  1.648444e-38  9.999966e-01  3.426560e-06  8.247054e-21 1.772701e-169
#> [334,]  2.651407e-37  9.999988e-01  1.152118e-06  1.032614e-21 4.460284e-172
#> [335,]  4.911658e-38  9.999978e-01  2.232609e-06  3.644541e-21 1.686708e-170
#> [336,]  1.289139e-37  9.999985e-01  1.528902e-06  1.770871e-21 2.109222e-171
#> [337,]  7.793898e-37  9.999992e-01  7.546689e-07  4.609737e-22 4.369434e-173
#> [338,]  7.446140e-37  9.999992e-01  7.683068e-07  4.769839e-22 4.820990e-173
#> [339,]  2.915634e-38  9.999973e-01  2.739577e-06  5.383391e-21 5.188459e-170
#> [340,]  4.103907e-36  9.999996e-01  3.932604e-07  1.330617e-22 1.219068e-174
#> [341,]  2.651407e-37  9.999988e-01  1.152118e-06  1.032614e-21 4.460284e-172
#> [342,]  1.653347e-39  9.999916e-01  8.447432e-06  4.605836e-20 2.514304e-167
#> [343,]  4.872912e-38  9.999978e-01  2.239558e-06  3.666195e-21 1.715736e-170
#> [344,]  3.412952e-38  9.999974e-01  2.575409e-06  4.785159e-21 3.695491e-170
#> [345,]  6.816158e-38  9.999980e-01  1.963240e-06  2.852313e-21 8.325691e-171
#> [346,]  3.791342e-38  9.999975e-01  2.471322e-06  4.423252e-21 2.946380e-170
#> [347,]  6.606534e-37  9.999992e-01  8.052329e-07  5.216351e-22 6.238530e-173
#> [348,]  6.816158e-38  9.999980e-01  1.963240e-06  2.852313e-21 8.325691e-171
#> [349,]  2.669222e-38  9.999972e-01  2.836158e-06  5.750950e-21 6.275704e-170
#> [350,]  6.816158e-38  9.999980e-01  1.963240e-06  2.852313e-21 8.325691e-171
#> [351,]  1.916511e-38  9.999968e-01  3.229851e-06  7.368063e-21 1.281287e-169
#> [352,]  3.454557e-38  9.999974e-01  2.563194e-06  4.741988e-21 3.600266e-170
#> [353,]  2.885203e-36  9.999995e-01  4.515667e-07  1.731847e-22 2.604488e-174
#> [354,]  1.289139e-37  9.999985e-01  1.528902e-06  1.770871e-21 2.109222e-171
#> [355,]  3.762826e-37  9.999990e-01  1.004249e-06  7.947253e-22 2.097910e-172
#> [356,]  2.960285e-37  9.999989e-01  1.103365e-06  9.509154e-22 3.517637e-172
#> [357,]  3.062136e-38  9.999973e-01  2.687382e-06  5.189558e-21 4.668363e-170
#> [358,]  3.749005e-39  9.999939e-01  6.126553e-06  2.496692e-20 4.308835e-168
#> [359,]  3.791342e-38  9.999975e-01  2.471322e-06  4.423252e-21 2.946380e-170
#> [360,]  9.242921e-38  9.999983e-01  1.742104e-06  2.271237e-21 4.319532e-171
#> [361,]  3.762826e-37  9.999990e-01  1.004249e-06  7.947253e-22 2.097910e-172
#> [362,]  9.242921e-38  9.999983e-01  1.742104e-06  2.271237e-21 4.319532e-171
#> [363,]  3.615296e-37  9.999990e-01  1.020133e-06  8.188599e-22 2.286716e-172
#> [364,]  3.791342e-38  9.999975e-01  2.471322e-06  4.423252e-21 2.946380e-170
#> [365,]  9.663583e-39  9.999958e-01  4.225362e-06  1.229649e-20 5.602169e-169
#> [366,]  5.299114e-39  9.999947e-01  5.348693e-06  1.927321e-20 2.044357e-168
#> [367,]  1.839916e-37  9.999987e-01  1.329715e-06  1.357143e-21 9.800438e-172
#> [368,]  1.143754e-38  9.999960e-01  3.954980e-06  1.084011e-20 3.896313e-169
#> [369,]  3.069486e-39  9.999934e-01  6.626647e-06  2.899525e-20 6.629552e-168
#> [370,]  3.046115e-38  9.999973e-01  2.692918e-06  5.209960e-21 4.721423e-170
#> [371,]  2.339504e-38  9.999970e-01  2.986745e-06  6.347012e-21 8.337478e-170
#> [372,]  8.505936e-38  9.999982e-01  1.799840e-06  2.416879e-21 5.166409e-171
#> [373,]  1.078011e-36  9.999993e-01  6.644819e-07  3.616677e-22 2.172274e-173
#> [374,]  2.528322e-39  9.999928e-01  7.150635e-06  3.352209e-20 1.006862e-167
#> [375,]  6.156720e-39  9.999950e-01  5.042966e-06  1.722766e-20 1.479774e-168
#> [376,]  2.651407e-37  9.999988e-01  1.152118e-06  1.032614e-21 4.460284e-172
#> [377,]  3.412952e-38  9.999974e-01  2.575409e-06  4.785159e-21 3.695491e-170
#> [378,]  7.818156e-38  9.999981e-01  1.860380e-06  2.574211e-21 6.195613e-171
#> [379,]  1.339101e-38  9.999963e-01  3.717700e-06  9.634112e-21 2.774010e-169
#> [380,]  1.443952e-39  9.999911e-01  8.908415e-06  5.096800e-20 3.366122e-167
#> [381,]  3.032785e-37  9.999989e-01  1.092939e-06  9.338605e-22 3.338955e-172
#> [382,]  1.289139e-37  9.999985e-01  1.528902e-06  1.770871e-21 2.109222e-171
#> [383,]  1.916511e-38  9.999968e-01  3.229851e-06  7.368063e-21 1.281287e-169
#> [384,]  3.311022e-38  9.999974e-01  2.606232e-06  4.894923e-21 3.944971e-170
#> [385,]  8.041095e-38  9.999982e-01  1.839969e-06  2.520639e-21 5.831432e-171
#> [386,]  5.279942e-38  9.999978e-01  2.170160e-06  3.452674e-21 1.443392e-170
#> 
#> $groups
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    0    0    0    0
#> [2,]    0    1    0    0    0
#> [3,]    0    0    1    0    0
#> [4,]    0    0    0    1    0
#> [5,]    0    0    0    0    1
#> 
#> $ngroups
#> [1] 5
#> 
#> $sigmasq
#> [1] 0.0001367895 0.0001367895 0.0001367895 0.0001367895 0.0001367895
#>